+++ frontend/src/metabase-lib/lib/utils/create-lookup-by-property.ts
