+++ frontend/src/metabase-lib/lib/utils/memoize-class.ts
