+++ frontend/src/metabase-lib/lib/utils/sort-object.ts
